Remove temporary directory usage in PathMapping repositories tests
